+++ docker/pytorch-aarch64/scripts/build-torchaudio.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 # *******************************************************************************
-# Copyright 2023-2024 Arm Limited and affiliates.
+# Copyright 2024 Arm Limited and affiliates.
 # SPDX-License-Identifier: Apache-2.0
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
@@ -20,25 +20,21 @@
 set -euo pipefail
 
 cd $PACKAGE_DIR
-readonly version=$TORCHVISION_VERSION
+readonly version=$TORCHAUDIO_VERSION
 readonly src_host=https://github.com/pytorch
-readonly src_repo=vision
+readonly src_repo=audio
 readonly num_cpus=$(nproc)
 
-# Clone TorchVision
+# Clone TorchAudio
 git clone ${src_host}/${src_repo}.git
 cd ${src_repo}
 git checkout v$version -b v$version
 git submodule sync
 git submodule update --init --recursive
 
-# Support for fixed format kernels
-patch -p1 < $PACKAGE_DIR/torchvision.patch
-rm $PACKAGE_DIR/torchvision.patch
-
-# Building TorchVision
+# Building TorchAudio
 CXX_FLAGS="${BASE_CFLAGS}" LD_FLAGS="${BASE_LDFLAGS}" FORCE_CUDA=0 python setup.py bdist_wheel
 
-# Installing the TorchVision python wheel via pip, dependencies are ignored as PyTorch is
+# Installing the TorchAudio python wheel via pip, dependencies are ignored as PyTorch is
 # a dependency, which would cause it to overwrite the locally build PyTorch
 pip install --no-deps $(ls -tr dist/*.whl | tail)